
\name{LogBF}
\alias{LogBF}
\title{Logarithm of a Bayes factor}
\usage{
LogBF(delta,X,y,n,a)
}
\arguments{
\item{delta}{indicatior variables}
\item{X}{matrix}
\item{y}{vector}
\item{n}{the number of observations}
\item{a}{constant}
}
\description{
Find the logarithm of a Bayes factor
}
\examples{
}
